requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-extensions==3.1.3
djangorestframework==3.12.4
idna==3.2
psycopg2==2.9.1
psycopg2-binary==2.9.1
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
psycopg2-binary==2.9.1
celery==5.1.2
redis==3.5.3
djangorestframework-api-key==2.0.0
pycodestyle==2.6.0